app/requirements.txt
llama-index
llama-index-llms-gemini
flask
python-dotenv
google-generativeai
setuptools
llama-index-vector-stores-mongodb